[util] Make zone functions 64-bit clean
This allows nq and qw clients/servers to use over 2GB of memory if
necessary.

+++ include/QF/zone.h
@@ -92,31 +92,32 @@
 
 typedef struct memzone_s memzone_t;
 
-void Memory_Init (void *buf, int size);
+void Memory_Init (void *buf, size_t size);
 
-void Z_ClearZone (memzone_t *zone, int size, int zone_offset, int ele_size);
+void Z_ClearZone (memzone_t *zone, size_t size, size_t zone_offset,
+				  size_t ele_size);
 void Z_Free (memzone_t *zone, void *ptr);
-void *Z_Malloc (memzone_t *zone, int size);			// returns 0 filled memory
-void *Z_TagMalloc (memzone_t *zone, int size, int tag);
-void *Z_Realloc (memzone_t *zone, void *ptr, int size);
+void *Z_Malloc (memzone_t *zone, size_t size);		// returns 0 filled memory
+void *Z_TagMalloc (memzone_t *zone, size_t size, int tag);
+void *Z_Realloc (memzone_t *zone, void *ptr, size_t size);
 void Z_Print (memzone_t *zone);
 void Z_CheckHeap (memzone_t *zone);
 void Z_SetError (memzone_t *zone, void (*err) (void *data, const char *msg),
 				 void *data);
-void Z_CheckPointer (const memzone_t *zone, const void *ptr, int size);
+void Z_CheckPointer (const memzone_t *zone, const void *ptr, size_t size);
 
 void Hunk_Print (qboolean all);
-void *Hunk_Alloc (int size);		// returns 0 filled memory
-void *Hunk_AllocName (int size, const char *name);
-int	Hunk_LowMark (void) __attribute__((pure));
-void Hunk_FreeToLowMark (int mark);
-void *Hunk_TempAlloc (int size);
+void *Hunk_Alloc (size_t size);		// returns 0 filled memory
+void *Hunk_AllocName (size_t size, const char *name);
+size_t Hunk_LowMark (void) __attribute__((pure));
+void Hunk_FreeToLowMark (size_t mark);
+void *Hunk_TempAlloc (size_t size);
 void Hunk_Check (void);
 
 
 
 struct cache_user_s;
-typedef void *(*cache_allocator_t) (struct cache_user_s *c, int size, const char *name);
+typedef void *(*cache_allocator_t) (struct cache_user_s *c, size_t size, const char *name);
 typedef void (*cache_loader_t) (void *object, cache_allocator_t allocator);
 typedef struct cache_user_s {
 	void	*data;
@@ -130,7 +131,7 @@ void *Cache_Check (cache_user_t *c);
 // if present, otherwise returns NULL
 
 void Cache_Free (cache_user_t *c);
-void *Cache_Alloc (cache_user_t *c, int size, const char *name);
+void *Cache_Alloc (cache_user_t *c, size_t size, const char *name);
 // Returns NULL if all purgable data was tossed and there still
 // wasn't enough room.
 void Cache_Report (void);
